Remove logger.debug validation on prevent_debugger
As @leafac and @carloslopes told, I am removing the logger.debug from black
list.

"I don't agree with the decition of not allowing logger.debug. It may be
useful to leave some debug level logs in production code. Then, when
some nasty bug arrises, the log level in production processes can be
changed for a while to track it down." @leafac

"Yes, logger.debug is totally acceptable and useful in production."
@carloslopes